ckanext/dcor_schemas/tests/requirements.txt
pytest
pytest-ckan
requests
requests_toolbelt